+++ src/utils/expansions_utils_2.c
@@ -1,0 +1,36 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   expansions_utils_2.c                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dyunta <dyunta@student.42madrid.com>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/11/29 19:49:40 by dyunta            #+#    #+#             */
+/*   Updated: 2024/11/29 19:50:02 by dyunta           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../include/minishell.h"
+
+/*
+ * find_var finds the correct value in t_list *env at key, allocates enough
+ * space for value and returns it.
+*/
+char	*find_var(const t_list *env, char *key, int errcode)
+{
+	t_var	*var;
+	char	*empty_str;
+
+	if (ft_strncmp(key, "?", ft_strlen(key)) == 0)
+		return (ft_itoa(errcode));
+	while (env)
+	{
+		var = ((t_var *)env->content);
+		if (ft_strlen(key) == ft_strlen(var->key)
+			&& ft_strncmp(key, var->key, ft_strlen(key)) == 0)
+			return (ft_strdup(var->value));
+		env = env->next;
+	}
+	empty_str = ft_calloc(1, 1);
+	return (empty_str);
+}
